+++ packages/ui/src/components/App.tsx
@@ -1,20 +1,18 @@
-import { FC, useState } from 'react';
-import Tabs from '@mui/material/Tabs';
-import Tab from '@mui/material/Tab';
+import { FC, useCallback, useState } from 'react';
 import { SpeciesTab } from './SpeciesTab';
 import { NewAdditionsTab } from './NewAdditionsTab';
-import { TaxonChangesTab } from './TaxonChangesTab';
+import { CuratedSpeciesDataMinified } from '@imerss/inat-curated-species-list-common';
 
 export interface AppProps {
   readonly speciesDataUrl: string;
   readonly curatorUsernames: string[];
   readonly placeId: number;
+  readonly showLastGeneratedDate: boolean;
   readonly showRowNumbers?: boolean;
   readonly showReviewerCount?: boolean;
   readonly newAdditionsDataUrl?: string;
   readonly showNewAdditions?: boolean;
   readonly showTaxonChanges?: boolean;
-  readonly taxonChangesDataUrl?: string;
   readonly lang?: any;
 }
 
@@ -26,29 +24,40 @@ export const App: FC<AppProps> = ({
   showReviewerCount,
   newAdditionsDataUrl,
   showNewAdditions,
-  showTaxonChanges,
-  taxonChangesDataUrl,
+  showLastGeneratedDate,
 }) => {
   const [tabIndex, setTabIndex] = useState(0);
-
-  const onChangeTab = (_e: React.SyntheticEvent | null, newValue: number) => {
-    setTabIndex(newValue);
-  };
-
+  const [lastGenerated, setLastGeneratedDate] = useState('');
   const hasNewAdditions = showNewAdditions && newAdditionsDataUrl;
-  const hasTaxonChanges = showTaxonChanges && taxonChangesDataUrl;
+
+  const onLoadSpeciesData = useCallback((data: CuratedSpeciesDataMinified) => {
+    setLastGeneratedDate(new Date(data.dateGenerated).toISOString().split('T')[0] as string);
+  }, []);
 
   const getTabs = () => {
-    if (!hasNewAdditions && !hasTaxonChanges) {
+    if (!hasNewAdditions) {
       return null;
     }
 
+    const speciesTabClass = tabIndex === 0 ? 'inat-curated-species-list-tab-selected' : '';
+    const newAdditionsTabClass = tabIndex === 1 ? 'inat-curated-species-list-tab-selected' : '';
+
     return (
-      <Tabs value={tabIndex} onChange={onChangeTab} className="inat-curated-species-list-tabs">
-        <Tab label="Species" disableRipple />
-        {hasNewAdditions && <Tab label="New Additions" disableRipple />}
-        {hasTaxonChanges && <Tab label="Taxon Changes" disableRipple />}
-      </Tabs>
+      <div style={{ position: 'relative' }}>
+        {showLastGeneratedDate && lastGenerated && (
+          <div className="inat-curated-species-list-last-generated-date">
+            Last generated: <span>{lastGenerated}</span>
+          </div>
+        )}
+        <ul className="inat-curated-species-list-tabs">
+          <li className={speciesTabClass} onClick={() => setTabIndex(0)}>
+            <button>Species</button>
+          </li>
+          <li className={newAdditionsTabClass} onClick={() => setTabIndex(1)}>
+            <button>New Additions</button>
+          </li>
+        </ul>
+      </div>
     );
   };
 
@@ -58,6 +67,7 @@ export const App: FC<AppProps> = ({
       <div style={{ display: tabIndex === 0 ? 'block' : 'none' }}>
         <SpeciesTab
           dataUrl={speciesDataUrl}
+          onLoad={onLoadSpeciesData}
           curatorUsernames={curatorUsernames}
           placeId={placeId}
           showRowNumbers={showRowNumbers}
@@ -68,12 +78,6 @@ export const App: FC<AppProps> = ({
       {hasNewAdditions && (
         <div style={{ display: tabIndex === 1 ? 'block' : 'none' }}>
           <NewAdditionsTab dataUrl={newAdditionsDataUrl} />
-        </div>
-      )}
-
-      {hasTaxonChanges && (
-        <div style={{ display: tabIndex === 2 ? 'block' : 'none' }}>
-          <TaxonChangesTab dataUrl={taxonChangesDataUrl} />
         </div>
       )}
     </>